+++ src/Compiler/TypeToPhpDoc.php
@@ -6,8 +6,10 @@ namespace Efabrica\PHPStanLatte\Compiler;
 
 use PHPStan\PhpDoc\TypeStringResolver;
 use PHPStan\Type\ErrorType;
+use PHPStan\Type\Generic\TemplateType;
 use PHPStan\Type\StaticType;
 use PHPStan\Type\Type;
+use PHPStan\Type\TypeTraverser;
 use PHPStan\Type\VerbosityLevel;
 use Throwable;
 
@@ -25,6 +27,14 @@ final class TypeToPhpDoc
         if ($type instanceof StaticType) {
             $type = $type->getStaticObjectType();
         }
+
+        // replace unresolved template types with their bounds (T of stdClass -> stdClass)
+        $type = TypeTraverser::map($type, static function (Type $type, callable $traverse): Type {
+            if ($type instanceof TemplateType) {
+                return $traverse($type->getBound());
+            }
+            return $traverse($type);
+        });
 
         $phpDoc = $type->describe(VerbosityLevel::precise());
         try {