+++ src/Template/Form/Form.php
@@ -75,6 +75,13 @@ final class Form implements NameTypeItem, ControlHolderInterface, JsonSerializab
         );
     }
 
+    public function withType(Type $type): self
+    {
+        $clone = clone $this;
+        $clone->type = $type;
+        return $clone;
+    }
+
     #[ReturnTypeWillChange]
     public function jsonSerialize()
     {