+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - stubs for filters accepting any value which could be converted to string
+- Resolving of generic template types in latte context
 
 ### Fixed
 - exit() and die() evaluated as early terminating call